Move attributes all to same package
