+++ src/main/java/seedu/address/logic/finance/parser/SpendCommandParser.java
@@ -13,13 +13,13 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.finance.commands.SpendCommand;
 import seedu.address.logic.finance.parser.exceptions.ParseException;
+import seedu.address.model.finance.attributes.Amount;
 import seedu.address.model.finance.attributes.Category;
+import seedu.address.model.finance.attributes.Description;
 import seedu.address.model.finance.attributes.Place;
+import seedu.address.model.finance.attributes.TransactionDate;
 import seedu.address.model.finance.attributes.TransactionMethod;
-import seedu.address.model.finance.logentry.Amount;
-import seedu.address.model.finance.logentry.Description;
 import seedu.address.model.finance.logentry.SpendLogEntry;
-import seedu.address.model.finance.logentry.TransactionDate;
 
 /**
  * Parses input arguments and creates a new SpendCommand object
