+++ src/main/java/seedu/address/model/finance/util/SampleDataUtil.java
@@ -6,14 +6,14 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.finance.FinanceLog;
 import seedu.address.model.finance.ReadOnlyFinanceLog;
+import seedu.address.model.finance.attributes.Amount;
 import seedu.address.model.finance.attributes.Category;
+import seedu.address.model.finance.attributes.Description;
 import seedu.address.model.finance.attributes.Place;
+import seedu.address.model.finance.attributes.TransactionDate;
 import seedu.address.model.finance.attributes.TransactionMethod;
-import seedu.address.model.finance.logentry.Amount;
-import seedu.address.model.finance.logentry.Description;
 import seedu.address.model.finance.logentry.LogEntry;
 import seedu.address.model.finance.logentry.SpendLogEntry;
-import seedu.address.model.finance.logentry.TransactionDate;
 
 /**
  * Contains utility methods for populating {@code FinanceLog} with sample data.
