+++ src/main/java/seedu/address/storage/finance/JsonAdaptedSpendLogEntry.java
@@ -9,14 +9,14 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.model.finance.attributes.Amount;
 import seedu.address.model.finance.attributes.Category;
+import seedu.address.model.finance.attributes.Description;
 import seedu.address.model.finance.attributes.Place;
+import seedu.address.model.finance.attributes.TransactionDate;
 import seedu.address.model.finance.attributes.TransactionMethod;
-import seedu.address.model.finance.logentry.Amount;
-import seedu.address.model.finance.logentry.Description;
 import seedu.address.model.finance.logentry.LogEntry;
 import seedu.address.model.finance.logentry.SpendLogEntry;
-import seedu.address.model.finance.logentry.TransactionDate;
 
 /**
  * Jackson-friendly version of {@link SpendLogEntry}.
